Offer details works

+++ src/main/java/ma/Ensate/StuRent/AddOffer/OfferRepository.java
@@ -2,5 +2,5 @@ package ma.ensate.sturent.AddOffer;
 
 import org.springframework.data.repository.CrudRepository;
 
-public interface OfferRepository extends CrudRepository<Offer, Integer> {
+public interface OfferRepository extends CrudRepository<Offer, Long> {
 }
